requirements.txt
# Core dependencies
biopython>=1.81
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0

# Protein Language Models
fair-esm>=2.0.0

# Data handling
requests>=2.31.0
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0

# Visualization
plotly>=5.14.0
umap-learn>=0.5.3

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
